fresh version of the code, all but one tests pass, finishing TODAYS (todo's for today)
