--- Maker/resources/decentral.contracts.md
@@ -1,37 +1,0 @@
-<h1 class="contract">
-   create
-</h1>
-### Parameters
-Input parameters:
-
-* `issuer` (name of the party issuing tokens)
-
-Implied parameters: 
-
-* `account_name` (name of the party invoking and signing the contract)
-
-### Intent
-INTENT. The intention of the author and the invoker of this contract is to print output. It shall have no other effect.
-
-### Term
-TERM. This Contract expires at the conclusion of code execution.
-
-<h1 class="contract">
-   issue
-</h1>
-### Parameters
-Input parameters:
-
-* `to` (name of the party the tokens are issued to)
-* `quantity` (the quantity of tokens issued)
-* `memo` (a message displayed when the action executed)
-
-Implied parameters: 
-
-* `account_name` (name of the party invoking and signing the contract)
-
-### Intent
-INTENT. The intention of the author and the invoker of this contract is to print output. It shall have no other effect.
-
-### Term
-TERM. This Contract expires at the conclusion of code execution.
