--- Maker/src/test.sh
@@ -1,14 +1,0 @@
-# propose
-# update price feeds
-# vote 
-# open
-# lock
-# draw
-# wipe 
-# shut
-# open
-# lock
-# draw
-# update 
-# liquify
-# settle
